+++ packages/measures/perf/agent-storm/fake-agent.ts
@@ -1,0 +1,39 @@
+import { dirname, join, resolve } from 'node:path'
+import { existsSync } from 'node:fs'
+import { createRequire } from 'node:module'
+import { fileURLToPath } from 'node:url'
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
+const require = createRequire(import.meta.url)
+
+export function buildFakeAgentScript(nodesPerAgent: number): object {
+    const actions: object[] = []
+    for (let i = 0; i < nodesPerAgent; i++) {
+        actions.push({
+            type: 'create_node',
+            title: `Perf Node ${i}`,
+            summary: `Synthetic node ${i} produced by perf-agent-storm.`,
+            content: `Node body for index ${i}. Generated by the perf harness.`,
+        })
+    }
+    actions.push({ type: 'exit', code: 0 })
+    return { actions }
+}
+
+export function buildAgentPrompt(script: object): string {
+    return `### FAKE_AGENT_SCRIPT ###\n${JSON.stringify(script)}\n### END_FAKE_AGENT_SCRIPT ###`
+}
+
+export function resolveFakeAgentEntrypoint(): { dir: string; entry: string } {
+    // measures/perf/agent-storm -> measures/perf -> measures -> packages -> repo root
+    const repoRoot = resolve(__dirname, '..', '..', '..', '..')
+    const dir = join(repoRoot, 'tools', 'vt-fake-agent')
+    const entry = join(dir, 'src', 'index.ts')
+    if (!existsSync(entry)) throw new Error(`vt-fake-agent entrypoint not found at ${entry}`)
+    return { dir, entry }
+}
+
+export function resolveTsxImportPath(): string {
+    return require.resolve('tsx')
+}
