+++ packages/measures/perf/agent-storm/filesystem.ts
@@ -1,0 +1,20 @@
+import { join } from 'node:path'
+import { readdirSync, statSync } from 'node:fs'
+
+export function countMarkdownFiles(dir: string): number {
+    let count = 0
+    const walk = (d: string): void => {
+        let entries: readonly string[]
+        try { entries = readdirSync(d) } catch { return }
+        for (const entry of entries) {
+            if (entry.startsWith('.')) continue
+            const full = join(d, entry)
+            let stat
+            try { stat = statSync(full) } catch { continue }
+            if (stat.isDirectory()) walk(full)
+            else if (stat.isFile() && entry.endsWith('.md')) count++
+        }
+    }
+    walk(dir)
+    return count
+}
